ok 90% website

+++ eBookShop/database/migrations/2021_05_18_162201_add_col_order_product.php
@@ -16,8 +16,8 @@ class AddColOrderProduct extends Migration
         Schema::table('orders', function (Blueprint $table) {
             $table->string('district');
             $table->string('note')->nullable();
-            $table->decimal('totalPrice',8,3);
-            $table->decimal('totalPriceFee',8,3);
+            $table->float('totalPrice');
+            $table->float('totalPriceFee');
             $table->integer('quantity');
         });
     }
